+++ src/components/CardSecurityCode.jsx
@@ -1,7 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import ErrorMessage from './common/ErrorMessage';
-import { Input, InputContainer, Span, Label, InputWrapper } from './common/styled';
+import { InputContainer, Span, Label, InputWrapper } from './common/styled';
 import QuestionContainer from './common/QuestionIcon';
+import Input from './common/Input';
 
 const MAX_CARD_CODE = 3;
 const CVC_EXPLANATION = `CVC번호는 카드뒷면의 7자리 숫자 중 뒷 3자리입니다.`;
